move to shared: chat/recipes/langs
